requirements.txt
asgiref==3.11.1
dj-database-url==3.1.0
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.1.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
qrcode==8.2
sqlparse==0.5.5
whitenoise==6.11.0
